hist_data_ingestor/requirements.txt
pydantic
sqlalchemy
typer
pytest
structlog
pandas
aiohttp
psycopg2-binary
tenacity
docker
python-dotenv
pandera
pyyaml